// @types/generated/contentful.d.ts
export interface ICardGridSectionFields {
  /** Title */
  title?: string | undefined;

  /** Subtitle */
  subtitle?: string | undefined;

  /** Cards */
  cards: (ICard | IBlogPost | IEvent | IPressArticle)[];
}

export interface ICardGridSection extends Entry<ICardGridSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "cardGridSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICtaSectionFields {
  /** Title */
  title: string;

  /** Body */
  body: string;

  /** Primary Button Text */
  primaryButtonText: string;

  /** Primary Button Link */
  primaryButtonLink: string;

  /** Primary Button open in New Window? */
  primaryIsExternal?: boolean | undefined;

  /** Secondary Button Text */
  secondaryButtonText?: string | undefined;

  /** Secondary Button Link */
  secondaryButtonLink?: string | undefined;

  /** Secondary open in New Window? */
  secondaryIsExternal?: boolean | undefined;
}

export interface ICtaSection extends Entry<ICtaSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "ctaSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IFullWidthImageSectionFields {
  /** Image */
  image: Asset;
}

export interface IFullWidthImageSection
  extends Entry<IFullWidthImageSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "fullWidthImageSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPageFields {
  /** Internal name */
  internalName: string;

  /** Page name */
  pageName?: string | undefined;

  /** Introduction */
  introduction?: string | undefined;

  /** Slug */
  slug: string;

  /** Sections */
  sections?:
    | (
        | ICardGridSection
        | ITwoColumnSection
        | ICtaSection
        | IFullWidthImageSection
      )[]
    | undefined;

  /** Body */
  body?: Document | undefined;
}

/** Page Template */

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "page";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ITwoColumnSectionFields {
  /** Title */
  title: string;

  /** Subtitle */
  subtitle?: string | undefined;

  /** Body */
  body: Document;
}

export interface ITwoColumnSection extends Entry<ITwoColumnSectionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "twoColumnSection";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}